Registry objects
This patch moves the registry management code into a new, separate
library.  We no longer hard-code Taconic's specific entities
(computations, connections, and value types).  Moreover, the
configuration for an entity can now be split into multiple files.  We're
going to use this in Taconic to allow, for instance, value types to be
defined across multiple libraries.  The base string type will be defined
in Taconic, for instance, while its Avro and Lua support will be added
in Palisades.