Simpler plugin descriptors
We're abandoning the complex tiered collection of entities and sections,
replacing it with a simple plugin architecture.  A plugin identifies a
shared library, and a “loader” function within that shared library.
Instead of expressing the complex extensible features in the filesystem,
they should be defined or registered in using application-specific
functions, which are called from the plugin library's loader function.