Load shared libraries
We now load in the shared libraries and loader functions specified in
each plugin descriptor file.  We don't need to worry about plugin
dependencies when loading in the shared libraries, since the linker
itself worries about link-time dependencies.  The plugin descriptor
dependencies only affect the order that we call the loader functions.
Which we're not doing yet.