Force privateer to link with libprivateer-tests
On some platforms (Ubuntu, for instance), the linker tries to outsmart
us — it notices that the privateer command doesn't use any of the
symbols in libprivateer-tests, and decides not to link with that
library.  In some of our test cases, though, we're going to try to load
in symbols from that library using a dlopen call with a NULL library
name.  This expects the symbols to already be present in the executable,
and so we really need the command to be linked with libprivateer-tests.

This patch adds an artificial (and unused) reference to one of the
symbols in libprivateer-tests, which should force the linker to always
include it.

Note that neither the privateer command, nor the libprivateer-tests
library, are installed; they're only used within the build directory in
our test cases.